refactor: Add user creation route and handler
This commit adds a new route and handler for creating a user. The route is `/users` and the handler is responsible for validating the user input and creating a new user in the database. This change enhances the functionality of the API by allowing users to be created through a dedicated endpoint.